+++ src/main/java/hu/tamas/university/repository/EventTypeRepository.java
@@ -3,7 +3,7 @@ package hu.tamas.university.repository;
 import hu.tamas.university.entity.EventType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventTypeRepository extends JpaRepository<EventType, Long> {
+public interface EventTypeRepository extends JpaRepository<EventType, Integer> {
 
 	EventType findEventTypeByType(String type);
 }
