+++ src/main/java/hu/tamas/university/repository/InvitationRepository.java
@@ -3,7 +3,7 @@ package hu.tamas.university.repository;
 import hu.tamas.university.entity.Invitation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvitationRepository extends JpaRepository<Invitation, Long> {
+public interface InvitationRepository extends JpaRepository<Invitation, Integer> {
 
 	Invitation findInvitationById(int id);
 }
