+++ src/main/java/hu/tamas/university/repository/LikesCommentRepository.java
@@ -3,7 +3,7 @@ package hu.tamas.university.repository;
 import hu.tamas.university.entity.LikesComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LikesCommentRepository extends JpaRepository<LikesComment, Long> {
+public interface LikesCommentRepository extends JpaRepository<LikesComment, Integer> {
 
 	LikesComment findLikesCommentById(int id);
