+++ src/main/java/hu/tamas/university/repository/LikesPostRepository.java
@@ -3,7 +3,7 @@ package hu.tamas.university.repository;
 import hu.tamas.university.entity.LikesPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LikesPostRepository extends JpaRepository<LikesPost, Long> {
+public interface LikesPostRepository extends JpaRepository<LikesPost, Integer> {
 
 	LikesPost findLikesPostById(int id);
 }
